backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
gspread>=6.0
google-auth>=2.28
google-api-python-client>=2.120
google-genai>=0.4.0
requests>=2.31
python-dotenv>=1.0
Pillow>=10.0